Changed SproxelConverter

+++ src/com/vloxlands/render/VoxelFace.java
@@ -17,10 +17,10 @@ public class VoxelFace
 	public Direction dir;
 	public Vector3f pos, tl, tr, bl, br;
 	
-	public long textureIndex;
+	public int textureIndex;
 	public int sizeX, sizeY, sizeZ;
 	
-	public VoxelFace(Direction dir, Vector3f pos, long texInd)
+	public VoxelFace(Direction dir, Vector3f pos, int texInd)
 	{
 		this(dir, pos, texInd, 1, 1, 1);
 	}
@@ -37,7 +37,7 @@ public class VoxelFace
 		updateVertices();
 	}
 	
-	public VoxelFace(Direction dir, Vector3f pos, long texInd, int sizeX, int sizeY, int sizeZ)
+	public VoxelFace(Direction dir, Vector3f pos, int texInd, int sizeX, int sizeY, int sizeZ)
 	{
 		super();
 		this.dir = dir;
@@ -127,8 +127,8 @@ public class VoxelFace
 	
 	public void render()
 	{
-		int texX = (int) textureIndex % 32;
-		int texY = (int) textureIndex / 32;
+		int texX = textureIndex % 32;
+		int texY = textureIndex / 32;
 		
 		glEnable(GL_CULL_FACE);
